+++ CHANGELOG.md
@@ -8,7 +8,7 @@ See this http://keepachangelog.com link for information on how we want this docu
 
 **Major Assistant & Tool Constructor Changes**
 
-Both Assistant & Tool have removed their (name, description, instructions) ordinal parameters in favor a single options object. 
+Both Assistant & Tool have removed their (name, description, instructions) ordinal parameters in favor a single options object. Now, the constructor signature is:
 
 ```javascript
 // Using Assistant.create() factory.
@@ -30,6 +30,10 @@ class MyAssistant extends Assistant {
   }
 })
 ```
+
+## Added
+
+A new `skipUpdate` option to use for deployments such as staging where it might be desirable to use the Assistants remote resource instructions or other properties.
 
 ## v1.1.0